add gate

+++ cold_start/ALDI.py
@@ -527,8 +527,9 @@ class ALDI(tf.keras.Model):
     # =====================
     def map_item(self, item_content, training=False):
         # ---------- FEATURE MASKING ----------
-        gate = tf.cast(tf.sigmoid(self.item_gate), item_content.dtype)
-        # gate = tf.sigmoid(self.item_gate)        # [content_dim]
+        # gate = tf.cast(tf.sigmoid(self.item_gate), item_content.dtype)
+        item_content = item_content.astype(np.float32)
+        gate = tf.sigmoid(self.item_gate)        # [content_dim]
         x = item_content * gate                  # element-wise
 
         for layer in self.item_layers:
